Editor: Fix the onClick handler of LegacyAlert
When the LegacyAlert component was lifted out, it was lifted out as if it was a
function within the renderer, rather than a component. As such, it was expecting
to receive the onClick handler directly, rather than a prop.

As it is a component which receives props, lets make it aware of that, and use
`props.migrateLegacy`, so the handler actually gets called.

Fixes #988.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>