api/flash: Remove support for external flasher
The option to support an external flasher when there's an internal default was
introduced as a workaround for a bug we have fixed long since. It was only
available on Windows, for two devices, and had to be opted into.

As the bug it was meant to work around is no more, lets drop this unnecessary
complexity, along with the win32 avrdude shipped with Chrysalis.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>