api/flash: Clear EEPROM before flashing
When flashing new firmware, the storage layout may change, and parts of it may
change owners. While we restore the previous contents in a structured manner, if
previously used space changes owner to a plugin that is new in the updated
firmware, it would start with random garbage in its storage slice.

To avoid this situation, clear the EEPROM before flashing, so when we restore,
we restore onto a fresh, empty state.

Fixes #934.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>