FirmwareAutoUpdate: Do not pretend an app update is available
The `useFirmwareAutoUpdate` hook was initially copied from `useAutoUpdate`, and
as part of the process, it set the `updateAvailable` state on the global
context. That's unrelated to firmware auto-updates, and the hook should not
touch it at all, so this patch removes that.

The result of the mistake was that whenever there was a firmware update,
Chrysalis' main sidebar also showed an "Apply update & restart" menu item by
mistake. With the hook not fiddling with state it has no business with, this is
no longer the case.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>