api/focus: Make waitForDFUBootloader actually perform a check
As long as this function existed, it was practically a no-op, it succeeded as
long as there was at least one USB device connected, of any kind. It did not
perform adequate checks to see if the requested VID/PID pair is among them.

This worked, because the flashing process rebooted the keyboard before checking
for the presence of a bootloader, so by the time we went to check, it was always
there anyway.

Things fell apart as soon as the function was supposed to do a real check. This
patch makes it do what it was always supposed to do: iterate over the devices
found, and only return success if the desired one is among them.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>