FirmwareUpdate: move defaultFirmwareFilename() back here
We need a `focusDeviceDescriptor` for this function, which we were pulling out
of `activeDevice.focus` previously. However, for keyboards already in bootloader
mode, there is no `activeDevice` to pull from.

FirmwareUpdate itself already has the descriptor via props, and doesn't need to
pull it from anywhere else, so move `defaultFirmwareFilename` back there, where
we have all the information we need at hand anyway.

This, along with the previous commit, fixes flashing when starting from
bootloader mode.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>