api/focus/keymap: Sort supported languages the same way we group them
We want to present the list of supported layouts grouped by their language (or
if that is unavailable, the language code, or group), and the groups should be
alphabetically sorted too.

To achieve this, we need to sort the list the same way we group them.

Addresses another issue identified in #649.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>