LayoutCard: Don't use an async function directly with useEffectOnce
`useEffectOnce`, just like `useEffect` itself, needs a plain, non-async
function. So instead of using doing initialization directly in the effect, push
it into an async function called _from_ the effect. The same pattern we use
elsewhere.

This gets rid of an error printed on the dev console when leaving the Layout
Cards screen.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>