CI: Merge the snapshot & release workflows
Having two workflows that do mostly the same thing, and the major difference
between them is whether they create a new release, or reuse an existing one,
when they should be sharing the vast majority of code, is rather pointless, and
leads to bugs.

Bugs like forgetting to add the settings required for Windows code signing to
the release workflow.

With the workflows merged, we have a single build workflow to maintain, share
the vast majority of code between snapshot and production builds, and thus, we
can avoid such mishaps in the future.

It also makes the release process slightly more reliable, because the snapshot
and release builds will not trip over each other, as there will be only one.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>