api/focus/keymap: withModifiers() should not apply mods to non-standard keys
Originally, `withModifiers()` was only called internally by the keymap db, so it
could safely assume that the list of keys it receives will fall within the
standard HID range, and can have modifiers applied safely.

However, with 62f6ba5034abcd31d71d38ff9a4f2f40be39f433, that is no longer the
case, and `withModifiers()` will be called with a keylist that can potentially
include already augmented keys.

It should only augment keycodes in the base HID range, and if it encounters a
keycode in the source list that is an augmented key, that should be ignored in
favour of the newly augmented ones.

Fixes #1026.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>