FirmwareUpdate: Change the confirmation dialog's confirm label to Continue
With the combination of our current flashing instructions, the "Cancel / Ok"
pair of buttons aren't exactly clear, they do not accurately convey what will
happen if either of them is pressed.

To make it clearer what they do, this changes the confirmation button from "Ok"
to "Continue", as that makes it clearer what pressing it will accomplish. The
"Cancel / Continue" pair should thus be clearer.

I intentionally avoided using "Upload" for the confirmation label, because
that'd just repeat what's on the button that popped this dialog up in the first
place.

Fixes #1023.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>