Keyboard104: Remove an unnecessary keyCode
The layout we used for Keyboard104 was a weird amalgamation of ISO and ANSI
layouts, and we ended up with three keys between `L` and `Return`, with the
closest one to `Return` being mapped to keycode 50, the alternate / non-us `#/~`
key.

However, none of the layouts shipped with CLDR map this anywhere, so it was a
constant `#/~` key there, and a source of confusion and frustration.

This patch removes it, and makes `Return` wider to accommodate the change.

There's another extra key in our layout compared to ANSI: between left shift and
`Z` - but that is heavily used by other layouts, so it's better to keep it, even
if it may not entirely be correct for US QWERTY.

Fixes #1044.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>