Preferences: SpaceCadet support
This adds support for toggling SpaceCadet on and off (and saving that setting),
and setting the global timeout. Requires the `SpaceCadetConfig` plugin to be
enabled (along with SpaceCadet, of course).

While both SpaceCadet and SpaceCadetConfig support a "no delay" mode, Chrysalis
does not, and will display the "no delay" mode the same as the normal active
mode.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>