tools/firmware-update: Always pull the latest tag, unless told otherwise
We no longer tie firmware and Chrysalis version numbers together, so the
firmware-update script will no longer attempt to figure out which version of the
firmware to download. It will simply download the latest, unless told otherwise.

To figure out the latest tag, the script does *not* use the GitHub REST API, but
checks the `location` header instead, because this does not require parsing JSON.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>