Implement a Breaking News component
The component is meant to be placed just below the application bar, on a screen
where it makes sense. It displays its children wrapped in an alert, and allows
the user to permanently dismiss the news item.

The component saves dismissals in a key among its settings, so future breaking
news will still be shown.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>