When using eeprom.erase, return after 10 seconds
When using `eeprom.erase` to restore factory settings, the keyboard will erase
the eeprom and reboot before sending us a reply. We can't await for that to
return, as it never will.

While we could - and should - fix the firmware, we still want to handle older
firmware, so we need a Chrysalis-side solution too. This solution is  to wait 10
seconds in `clearEEPROM()`, and just return. Ten seconds should be enough to
clear the eeprom, even if we're using a pretty large emulated storage where
writes are slow. They're not *this* slow.

With this, the factory reset in Preferences will eventually complete. Since we
enforce a reboot afterwards, this factory reset will now disconnect from the
keyboard as well, and land us back on the Keyboard Selection screen.

Fixes #1056.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>