Perform some sanity checks during startup
When starting up, perform some sanity checks, to verify that Chrysalis is
running in an environment that allows it to function correctly. Among other
things, we verify that the Electron version we're running against is the
expected one, and that static files bundled with the package are accessible.

If any of the sanity checks fail, a message is displayed, urging the user to use
an official build. The user can still navigate away, if they choose to do so,
the message is only shown once on startup.

This is mostly here to combat the `chrysalis` package in AUR, which is causing a
lot of frustration and support requests.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>