Editor: Do not disable saving if LayerNames is not in use
In case the `LayerNames` plugin is not enabled firmware-side, getting the stored
size of the layer names still returns a non-zero value. We need to check if
there is *any* storage size, before comparing the serialized length, otherwise
we mistakenly disable saving (and don't even show a warning, because the warning
correctly checks for the presence of the plugin).

Fixes #1119.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>