styles/keymap.css: Disable user-select for the keymap SVGs
Being able to select text is in general a desirable feature. However, in case of
the keymap, being able to do so is distracting, annoying, and not useful at all.
This little change disables that, preventing one from selecting the *text* of
key labels. It does not prevent selecting the buttons themselves, of course.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>