FirmwareUpdate: Rework how we do factory resets
We used to do factory resets *before* flashing, because doing so was easy.
However, doing so is wrong, because that does not let us perform the feat when
started from bootloader mode. We want to be able to do that, too.

As such, we'll do the factory reset after flashing, right before restoring
EEPROM.

While doing the rearrangement, I also fixed an issue that prevented us from
removing `Model100Bootloader`: we did not correctly remember if we started from
bootloader mode. Now we do, so the extra descriptor is no longer necessary.

Fixes #1112.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>