ActiveDevice: Improve `focusDetected()`
First of all, this does not need to be async. Second, if the connected keyboard
is not in application mode, then it doesn't support focus either, so return that
early.

Adjust `.supportFocusCommand()`, because due to the previous commits,
`this.supported_commands()` may now return undefined. In that case, return
false.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>