Preferences: Fix a possible case of uncaught exception
When rendering the selector for "Default layer", we require the `keymap`
variable to be in a good state. We're initializing it in `initialize()` (used
via a `useEffect()`), and we render things when `loaded` becomes true.

However, the selector's menu items were *not* guarded by `loaded`, so there was
a small race where React would attempt to re-render, before `initialize()`
completing, and then `keymap` would still be in an unknown state. This sometimes
led to white screens and a renderer crash.

Lets guard this part of the code properly, and only try to render the menu items
once *everything* is loaded.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>