Toast: Force messages into a string
There are certain points within Chrysalis where we send non-string messages to
Toast. While ideally, we shouldn't be doing that, we do, and fixing that would
be a larger undertaking.

As an temporary - and as a robustness - measure, force messages into a string,
so we can display them even if they were some other type originally. The main
reason we want a string is that if Toast receives an Error object, that is
something React will not be able to render, and will complain about a component
being the wrong type, and reward us with an uncaught exception and a white
screen.

By forcing messages into a string, we avoid that. This is more robust, and helps
us diagnose problems better, too, because we'll have the error displayed.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>