Toast: Log the messages received
It is sometimes useful to know what toasts were displayed in a session, to be
able to more accurately replay what transpired. For this reason, log any toast
messages received, at debug level, with a custom label for easy filtering.

Do not log progress updates, however, we're not interested in those.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>