Editor: Always show the "Copy layer" and "Paste layer" buttons
The Overview footer was only displayed in the past when we had empty layers. The
"Copy layer" and "Paste layer" buttons were later added to the same footer - but
the condition on displaying the footer remained intact. This resulted in these
buttons disappearing when we had no empty layers.

Lets display the footer unconditionally, and only do the hiding logic for the
empty layer toggle button.

Fixes #1162.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>