main/ipc_firmware: Only pull pre-releases when running -snapshot
To match user expectation (and the behaviour of Chrysalis auto-updates), we
should not consider firmware pre-releases as candidates, unless Chrysalis is
also a snapshot version.

To this end, if the Chrysalis version does not have a `-snapshot` tag, filter
out firmware pre-releases when trying to find the latest version.

Fixes #1163.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>