CI: Add a workflow to build & publish unsigned pull request artifacts
Lately, we have been seeing - and creating - pull requests where user testing
and feedback would be of incredible value. However, building from source is a
barrier we don't want to put in front of potential testers.

For this reason, lets build and publish artifacts for pull requests aswell. They
will not be signed - because we have no control over what gets built, really -,
and they will not be uploaded to GitHub. They are made available nevertheless,
and the workflow is set up so that it posts the URLs after the build to the
originating pull request.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>