Preferences: Make it possible to turn off chunked writes
Chrysalis has been using chunked writes on macOS since about forever, and we
switched to using those globally with the launch of the Model 100, as a possible
remedy for communication issues (it did not help much).

Since then, we fixed the Model100 firmware, and we don't really support the
buggy firmware in any shape or form, except for connecting and eventually
landing on the Firmware Update screen. As such, half the reason for using
chunked writes is gone.

Chunked writes, however, are slow. Very slow. It's not a great user experience.
We also discovered that even on macOS, they may not be needed anymore.

However, we do not want to just disable them, and hope for the best, so for now,
offer it as an option, hidden in Preferences / Developer Tools. There, because
we don't want people to fiddle with it without care. It's mostly there for
developers, to make it easier to test turning chunked writes off, without having
to recompile and re-launch Chrysalis.

This is a strictly temporary option, which will go away as soon as we learned
whether to keep chunked writes, or remove them completely.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>