Editor: Layer copy & paste should not depend on Colormap
It is not required for a device to have any colormap-related functionality,
whether it has LEDs or not. Copying and pasting layers should function
regardless of the presence of colormaps.

As such, make `copyLayer()` handle the case of a missing colormap correctly, and
do not require colormap for `hasCopiedLayer()`.

With these fixes, it is now possible to copy & paste layers on an Atreus.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>