hooks/usePluginEffect: Pass activeDevice to `initialize()`
When using `usePluginEffect()`, the most common case is to pull data from the
keyboard during initialization. To make that easier, the hook was passing
`activeDevice.focus` as the first argument.

However, we're moving towards using `activeDevice`, and many of the users of
this plugin would now want that, instead of just `focus`. To not break
compatibility (yet), pass `activeDevice` as the second argument.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>