api/focus: Try to guard the serialport opening
We have seen cases where Chrysalis reported uncaught exceptions occurring during
serialport opening, within a Promise in node-serialport. To try and work around
that, ask SerialPort to not auto-open the port, and we do that separately, so we
can wait on it (thus avoiding the exception happening in an un-awaited Promise),
and guard that with a try/catch block.

We'll still throw an exception, but hopefully we'll have more context this way.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>