Editor/Sidebar: Remove the auto-close/auto-open behaviour
The driving logic behind the auto-close/auto-open behaviour of collapsible
sections turned out to be more confusing than useful. Implementing a behaviour
that is intuitive and not confusing, is bordering the impossible, so lets not
try that.

This patch removes the functionality altogether: collapsible sections will start
off being closed, but once opened manually, they will stay open.

Fixes #1159.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>