ci: Only run the publish step for PRs when the source repo is ours
Because the publish step requires secrets that are unavailable for pull
requests, we can't - and don't want to, either - run the publish step, because
it would fail. As such, skip the step instead.

This does reduce the usefulness of building artifacts for each PR, but they
remain useful for PRs originating from the Chrysalis repo, so it is still worth
keeping.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>