Fix the colormap importing via backup & restore
When importing the colormap, we're importing both the palette, and the colormap,
but we were importing them in two steps, both of which were asynchronously
setting the same state. This resulted in the two async state changes stomping on
each others feet, and corrupting the colormap.

To remedy this issue, add a new helper that can set them both at the same time,
and thus, we avoid the race condition.

Fixes #1059.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>