Do not offer snapshot firmware on release versions of Chrysalis
Even though Chrysalis does not download snapshot firmware if Chrysalis itself
isn't a snapshot, it still offered *previously* downloaded snapshot firmware on
the Firmware Update screen.

To remedy this issue, the backend will now check the version of any downloaded
firmware, and skip over it if its a snapshot version while Chrysalis is not.

Fixes #1195.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>