FirmwareUpdate: Default to factory reset when starting from bootloader mode
When starting a firmware update from bootloader mode, we weren't doing an eeprom
erase before, which could lead to surprising results if people flashed firmware
from bootloader mode for one reason or another, and the eeprom layout changed.

To remedy that, lets default to factory reset when the keyboard is already in
bootloader mode. This avoids surprises in most cases, but still allows one to
turn it off, and just do a plain flash.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>