FirmwareUpdate: Display a warning when we find a keyboard in bootloader mode
Flashing when we start from bootloader mode has some drawbacks, like not being
able to migrate eeprom. The previous commit made Chrysalis default to enabling
factory reset in this case. This one displays a warning in place of the firmware
version (which is meaningless in the bootloader case anyway). On top of that,
Chrysalis will display an additional button that makes it easier to back out and
disconnect.

Thanks to @tlyu for the warning message, and for reviewing the UI.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>