Move EEPROM restore to layout backup screen
The layout backup already contains EEPROM settings.
Instead of having separate buttons for full backup/restore
in the preferences, this just extends the current restore
functionality with the ability to also restore the rest of
the EEPROM settings. The option to only restore layout
related settings is still available through a drop down.

When EEPROM settings are loaded, a table is shown with the
old and new value for each setting. Settings with unchanged
values are omitted, to reduce noise.

Signed-off-by: thehenkan <thehenkan@gmail.com>